+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Added `Summon Harem` ability to `Isekai Protagonist` progression.
+- Added `Playful Darkness` to `Summon Calamity` Overpowered ability.
+- Added `Signature Ability` to `Isekai Protagonist` progression. This can be chosen instead of `Signature Attack`.
+- Added `Summon Harem` ability to `Isekai Protagonist` progression. This ability replaces `Harem Magnet`.
 - Added `Channel Energy` selection to `God Emperor` progression.
 - Added `Divine Array` feature to `God Emperor` progression.
 - Added bard, ranger, paladin, and inquisitor spells into the `Isekai Protagonist` spellbook.